+++ CHANGELOG.md
@@ -4,7 +4,20 @@
 
 ### Added
 
-* Operation to remove asset pair
+* `RemoveAssetPairOp` operation to remove asset pair
+* `InitiateKYCRecoveryOpFrame`
+* `CreateKYCRecoveryRequestOpFrame`
+* `ReviewKYCRecoveryOpFrame`
+
+### Fixed
+
+* Tasks handling on `ChangeRoleRequest` reject
+
+## 3.4.1
+
+### Fixed
+
+* exception on ledger version upgrade
 
 ## 3.4.0